FEAT: Login & Bug Fixed & Logging front
