+++ public/index.html
@@ -18,7 +18,7 @@
       생략
     -->
     <title>Music Recommendation Recsys 07</title>
-    <script src="https://accounts.google.com/gsi/client" async defer></script>
+    <script src="https://accounts.google.com/gsi/client" defer></script>
   </head>
   <body>
     <div id="root"></div>
