+++ src/components/body.js
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React, { useState, useEffect } from "react";
 import Header from "./header";
 import Navigation from "./nav";
 import MiniPlayer from "./miniPlayer";
@@ -11,6 +11,8 @@ import Input from "./start_input/input"
 import HandleUrl from "./start_input/handleUrl"
 // import Select from "./start_input/select"
 import Search from "./start_input/search";
+import Login from "./login";
+import { Credentials } from "./spotifyAPI/Credentials";
 
 import CuratorPlayList from "./curatorPlayList";
 import axios from "axios";
@@ -18,11 +20,33 @@ import GoogleButton from "./googlebutton";
 
 const Body = (e) => {
   const [selectedCurator, setSelectedCurator] = useState("")
+  const [token, setToken] = useState('');
+
+  const spotify = Credentials();
+  useEffect(() => {
+    const fetchData = async () => {
+      const tokenResponse = await axios('https://accounts.spotify.com/api/token', {
+        headers: {
+          'Content-Type': 'application/x-www-form-urlencoded',
+          Authorization:
+            'Basic ' + btoa(spotify.ClientId + ':' + spotify.ClientSecret),
+        },
+        data: 'grant_type=client_credentials',
+        method: 'POST',
+      });
+
+      setToken(tokenResponse.data.access_token);
+    };
+
+    fetchData();
+    setInterval(fetchData, 3600 * 1000);  // run the code every hour
+  }, []);
 
   const handleCuratorClick = (val) => {
     setSelectedCurator(val);
   }
 
+  window.userEmail = 'NOT_LOGINED'
   return (
       <div>
         <header>
@@ -43,7 +67,9 @@ const Body = (e) => {
           {/* 기초 세팅 */}
           <Header />
           <Navigation />
-          <GoogleButton />
+          {/* <GoogleButton /> */}
+          <Login />
+          
           {/* <MiniPlayer /> */}
           <div className='dim'></div>
           <Player />
@@ -54,8 +80,8 @@ const Body = (e) => {
 
           {/* Start Input 페이지 */}
           <Input />
-          <HandleUrl />
-          <Search />
+          <HandleUrl token={token}/>
+          <Search token={token}/>
           {/* <Select /> */}
       
           {/* 큐레이터 전환 */}
